ivy: Refactor power config
Move the power config out of the main init scripts so that we can
be more versatile as to when we apply power configurations.

This also allows us to remove one configuration script from the
build system.

The only functional change is enabling zRAM at post-fs instead of fs,
to avoid a race condition between the mounting of the partitions in
the fstab and the enabling of zRAM.

Change-Id: Ia6888aaa55d6168706b64b252845487e323519bc