Update patch set 3
Patch Set 3:

I propose to already merge this like other devices did. It's fonctionnaly the same that your commits Evgeny, except the common config is not common. 
Then we should use your commits for a common bluetooth config as soon as it's merged in kitakami-common.
Are you ok with that?

Patch-set: 3
Label: Verified=0